requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-appconf==1.0.6
idna==3.10
packaging==24.1
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.3
